requirements.txt
antiorm==1.2.1
appdirs==1.4.3
asn1crypto==0.24.0
attrs==17.3.0
Automat==0.6.0
awsebcli==3.13.0
beautifulsoup4==4.6.0
bleach==2.1.1
botocore==1.10.29
brewer2mpl==1.4.1
cached-property==1.3.1
cement==2.8.2
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
colorama==0.3.7
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.1
cycler==0.10.0
Cython==0.27.3
db==0.1.1
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.7
django-adminlte2==0.2.3
django-multiselectfield==0.1.8
djangorestframework==3.7.3
docker==2.6.1
docker-pycreds==0.2.3
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fibonacci-heap-mod==1.0
Flask==0.12.2
Flask-API==1.0
gevent==1.2.2
greenlet==0.4.12
html5lib==1.0b10
httplib2==0.10.3
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.3
isodate==0.5.4
itsdangerous==0.24
jdcal==1.3
jedi==0.11.0
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
lxml==4.0.0
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.7.4
mpmath==1.0.0
mysqlclient==1.3.12
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
notebook==5.2.0
numpy==1.13.3
oauth2client==4.1.2
olefile==0.44
openpyxl==2.5.0
pandas==0.21.0
pandocfilters==1.4.2
parsel==1.2.0
parso==0.1.0
pathspec==0.5.5
pbkdf2==1.3
pickleshare==0.7.4
Pillow==4.3.0
pipenv==2018.5.18
plotly==2.1.0
prompt-toolkit==1.0.15
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyExecJS==1.5.0
Pygments==2.2.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
queuelib==1.4.2
requests==2.9.1
requests-toolbelt==0.8.0
rsa==3.4.2
scikit-learn==0.19.1
scipy==1.0.0
Scrapy==1.4.0
seaborn==0.8.1
semantic-version==2.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
sympy==1.1.1
tabulate==0.7.5
termcolor==1.1.0
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
Twisted==17.9.0
uritemplate==3.0.0
urllib3==1.22
utils==0.9.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
w3lib==1.18.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.48.0
websockets==3.4
Werkzeug==0.12.2
widgetsnbextension==3.0.6
wstools==0.4.5
xlrd==1.1.0
zeep==2.4.0
zope.interface==4.4.3
